Extract expression parsing to Util file
